[feat]サイドバーの実装、細かい調整
