--- app/(default)/chatInterface/index.module.css
@@ -1,181 +1,0 @@
-.heroContainer {
-  width: 70vw;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  text-align: center;
-  align-items: center;
-}
-
-.heroContainer.expanded {
-  margin: 4rem 0 4rem 0;
-}
-
-.hero {
-  width: 100%;
-  max-width: 900px;
-}
-
-.title {
-  font-size: 36px;
-  font-weight: bold;
-  color: #333;
-  margin-bottom: 10px;
-}
-
-.subtitle {
-  font-size: 27px;
-  font-weight: bold;
-  color: #6f627a;
-  margin-bottom: 50px;
-}
-
-.components {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-  box-sizing: border-box;
-  flex-direction: column;
-}
-
-.hideComponent {
-  display: flex;
-  flex-direction: column;
-  text-align: left;
-  box-sizing: border-box;
-}
-
-.hideComponent p {
-  font-weight: bold;
-  font-size: 20px;
-  color: #333;
-  margin: 1.5rem 0 0.3rem 0;
-}
-
-.form {
-  margin-bottom: 20px;
-}
-
-.mb {
-  margin-bottom: 20px;
-}
-
-.label {
-  display: block;
-  margin-bottom: 8px;
-}
-
-.input {
-  width: 100%;
-  padding: 8px;
-  box-sizing: border-box;
-}
-
-.buttonContainer {
-  display: flex;
-  gap: 10px;
-}
-
-.gearButton {
-  background: none;
-  border: none;
-  font-size: 24px;
-  cursor: pointer;
-}
-
-.button {
-  padding: 10px 20px;
-  border: none;
-  background-color: #0070f3;
-  color: white;
-  cursor: pointer;
-  border-radius: 4px;
-}
-
-.bgRed {
-  background-color: #ffeded;
-}
-
-.borderRed {
-  border: 1px solid #ff0000;
-}
-
-.textRed {
-  color: #ff0000;
-}
-
-.px {
-  padding-left: 16px;
-  padding-right: 16px;
-}
-
-.py {
-  padding-top: 16px;
-  padding-bottom: 16px;
-}
-
-.rounded {
-  border-radius: 8px;
-}
-
-.bold {
-  font-weight: bold;
-}
-
-.mb {
-  margin-bottom: 20px;
-}
-
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-
-.fadeIn {
-  animation: fadeIn 1.5s ease-in-out;
-}
-
-@keyframes slideDown {
-  from {
-    opacity: 0;
-    transform: translateY(-1rem);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0px);
-  }
-}
-
-.slideDown {
-  animation: slideDown 0.5s ease-in-out;
-}
-
-@keyframes slideUp {
-  from {
-    opacity: 1;
-    transform: translateY(0px);
-  }
-  to {
-    opacity: 0;
-    transform: translateY(-1rem);
-  }
-}
-
-.slideUp {
-  animation: slideUp 0.5s ease-in-out;
-}
-
-.hidden {
-  display: none;
-}
-
-.scoreText {
-  font-size: 24px;
-  font-weight: bold;
-  color: #333;
-}
