+++ app/(default)/page.module.css
@@ -1,0 +1,172 @@
+.heroContainer {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  text-align: center;
+  align-items: center;
+  transition: width 0.3s ease;
+}
+
+.heroContainer.sidebarOpen {
+  width: 60%;
+}
+
+.title {
+  font-size: 36px;
+  font-weight: bold;
+  color: #333;
+  margin-bottom: 10px;
+}
+
+.subtitle {
+  font-size: 27px;
+  font-weight: bold;
+  color: #6f627a;
+  margin-bottom: 50px;
+}
+
+.componentTop {
+  display: flex;
+  justify-content: center;
+  width: 100%;
+  box-sizing: border-box;
+  flex-direction: column;
+  padding-left: 60px;
+  padding-right: 60px;
+}
+
+.componentBottom {
+  display: flex;
+  justify-content: center;
+  width: 100%;
+  box-sizing: border-box;
+  flex-direction: column;
+}
+
+.resultBox {
+  margin-top: 60px;
+}
+
+.sidebar {
+  width: 0;
+  height: 100%;
+  background-color: #f0f0f0;
+  transition: width 0.3s ease;
+  overflow: hidden;
+}
+
+.sidebarOpen {
+  width: 40%;
+}
+
+.container {
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+  height: calc(100vh - 60px);
+  overflow: hidden;
+  position: relative;
+}
+
+.hideComponent {
+  display: flex;
+  flex-direction: column;
+  text-align: left;
+  box-sizing: border-box;
+}
+
+.hideComponent p {
+  font-weight: bold;
+  font-size: 20px;
+  color: #333;
+  margin: 1.5rem 0 0.3rem 0;
+}
+
+.bgRed {
+  background-color: #ffeded;
+}
+
+.borderRed {
+  border: 1px solid #ff0000;
+}
+
+.textRed {
+  color: #ff0000;
+}
+
+.px {
+  padding-left: 16px;
+  padding-right: 16px;
+}
+
+.py {
+  padding-top: 16px;
+  padding-bottom: 16px;
+}
+
+.rounded {
+  border-radius: 8px;
+}
+
+.bold {
+  font-weight: bold;
+}
+
+.fadeIn {
+  animation: fadeIn 1.5s ease-in-out;
+}
+
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
+}
+
+@keyframes slideDown {
+  from {
+    opacity: 0;
+    transform: translateY(-1rem);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0px);
+  }
+}
+
+.slideDown {
+  animation: slideDown 0.5s ease-in-out;
+}
+
+@keyframes slideUp {
+  from {
+    opacity: 1;
+    transform: translateY(0px);
+  }
+  to {
+    opacity: 0;
+    transform: translateY(-1rem);
+  }
+}
+
+.slideUp {
+  animation: slideUp 0.5s ease-in-out;
+}
+
+.hidden {
+  display: none;
+}
+
+.scoreText {
+  font-size: 24px;
+  font-weight: bold;
+  color: #333;
+}
+
+.mb {
+  margin-bottom: 20px;
+}
