+++ app/components/layouts/Footer/Footer.module.css
@@ -1,0 +1,25 @@
+.footer {
+  width: 100%;
+  height: 60px;
+  background-color: #f8f9fa;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  position: relative;
+  box-shadow: 0 -2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.text {
+  font-size: 14px;
+  color: #6c757d;
+}
+
+@media (max-width: 768px) {
+  .footer {
+    height: 50px;
+  }
+
+  .text {
+    font-size: 12px;
+  }
+}
