+++ app/components/layouts/Footer/Footer.tsx
@@ -1,5 +1,10 @@
 import React from 'react'
+import styles from './Footer.module.css'
 
 export default function Footer() {
-  return <></>
+  return (
+    <footer className={styles.footer}>
+      <p className={styles.text}>© 2024 Your Website. All rights reserved.</p>
+    </footer>
+  )
 }
