+++ app/components/layouts/Header/Header.tsx
@@ -5,13 +5,13 @@ import React from 'react'
 import { BiMenuAltLeft } from 'react-icons/bi'
 import styles from './Header.module.css'
 
-export default function Header() {
+export default function Header({ onMenuClick }: { onMenuClick: () => void }) {
   // const { data: session } = useSession()
 
   return (
     <header className={styles.header}>
-      <BiMenuAltLeft className={styles.icon} size={50} />
-      <p className={styles.text}>GPT-Logprobs</p>
+      <BiMenuAltLeft className={styles.icon} size={50} onClick={onMenuClick} />
+      <p className={styles.text}>Prompt Lab</p>
       {/* <nav className={styles.nav}>
         <ul className={styles.navList}>
           {session ? (
