+++ app/types/index.tsx
@@ -79,7 +79,6 @@ export interface InputBoxProps {
   isToggled: boolean
   setIsToggled: (value: boolean) => void
   isLoading: boolean
-  setIsTextareaExpanded: (value: boolean) => void
 }
 
 export interface CorrectBoxProps {
